api-python/requirements.txt
aniso8601==8.1.1
bcrypt==3.2.0
Cerberus==1.3.2
cffi==1.14.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.22.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.3
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
waitress==1.4.4
Werkzeug==1.0.1
WTForms==2.3.3
